+++ src/components/mypage/habits/HabitItem.tsx
@@ -1,18 +1,13 @@
 import { useState } from 'react';
 import { Habit, UserPoint } from '@prisma/client';
-import { useAddPointMutation } from '@/lib/mutations/useAddPointMutation';
 import HabitForm from './HabitForm';
 import {
   getCooldownStatus,
   getCurrentDayStatus,
 } from '@/lib/utils/habit.utils';
-import {
-  useDeleteHabitMutation,
-  useUpdateHabitMutation,
-} from '@/lib/mutations/useHabitMutation';
-import { toast } from '@/hooks/use-toast';
 import { ICONBUTTON_MODE } from '@/constants/mode.constants';
 import IconButton from '@/components/common/button/IconButton';
+import { useHabitItemHandlers } from '@/lib/hooks/useHabitItemHandlers';
 
 type HabitItemProps = {
   habit: Habit & { userPoints: UserPoint[] };
@@ -21,50 +16,16 @@ type HabitItemProps = {
 
 const HabitItem = ({ habit, userId }: HabitItemProps) => {
   const [isEditing, setIsEditing] = useState(false);
-  const addPointMutation = useAddPointMutation(userId);
-  const updateMutation = useUpdateHabitMutation(userId, habit.id);
-  const deleteMutation = useDeleteHabitMutation(userId, habit.id);
-
-  const handleAddPoint = () => {
-    addPointMutation.mutate(habit.id, {
-      onSuccess: () =>
-        toast({ title: '성공', description: '포인트가 추가되었습니다.' }),
-      onError: (err) =>
-        toast({
-          title: '실패',
-          description: `포인트 추가 실패: ${err.message}`,
-        }),
+  const { handleAddPoint, handleUpdateHabit, handleDeleteHabit, isAddPending } =
+    useHabitItemHandlers({
+      userId,
+      habitId: habit.id,
+      onEditToggle: setIsEditing,
     });
-  };
-
-  const handleUpdateHabit = (
-    updatedHabit: Omit<Habit, 'userId' | 'createdAt' | 'userPoints'>,
-  ) => {
-    updateMutation.mutate(updatedHabit, {
-      onSuccess: () =>
-        toast({ title: '성공', description: '습관이 수정되었습니다.' }),
-      onError: (err) =>
-        toast({ title: '실패', description: `습관 수정 실패: ${err.message}` }),
-    });
-    setIsEditing(false);
-  };
-
-  const handleDeleteHabit = () => {
-    deleteMutation.mutate(undefined, {
-      onSuccess: () =>
-        toast({ title: '성공', description: 'Habit이 삭제되었습니다.' }),
-      onError: (err) =>
-        toast({
-          title: '실패',
-          description: `Habit 삭제 실패: ${err.message}`,
-        }),
-    });
-  };
 
   const isValidDay = getCurrentDayStatus(habit);
   const isCooldownActive = getCooldownStatus(habit.userPoints);
-  const isDisabled =
-    !isValidDay || isCooldownActive || addPointMutation.isPending;
+  const isDisabled = !isValidDay || isCooldownActive || isAddPending;
 
   return (
     <div className="flex flex-col gap-2">
